refactor edit address space, review and configuration components
