+++ console/console-init/ui/src/components/TablePagination/TablePagination.tsx
@@ -5,12 +5,9 @@
 
 import React, { useCallback } from "react";
 import { useHistory, useLocation } from "react-router";
-import {
-  Pagination as TablePagination,
-  PaginationProps
-} from "@patternfly/react-core";
+import { Pagination, PaginationProps } from "@patternfly/react-core";
 
-export const Pagination: React.FC<PaginationProps> = ({
+export const TablePagination: React.FC<PaginationProps> = ({
   page,
   perPage,
   itemCount,
@@ -50,7 +47,7 @@ export const Pagination: React.FC<PaginationProps> = ({
 
   if (itemCount && itemCount > 0) {
     return (
-      <TablePagination
+      <Pagination
         itemCount={itemCount}
         perPage={perPage}
         page={page}
