+++ console/console-init/ui/src/modules/address-space/components/AddressSpaceConfiguration/AddressSpaceConfiguration.tsx
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-import * as React from "react";
+import React from "react";
 import {
   Grid,
   GridItem,
@@ -16,205 +16,64 @@ import {
   DropdownPosition,
   Radio
 } from "@patternfly/react-core";
-import { useQuery } from "@apollo/react-hooks";
 import { IDropdownOption } from "components/common/FilterDropdown";
-import {
-  RETURN_ADDRESS_SPACE_PLANS,
-  RETURN_NAMESPACES,
-  RETURN_AUTHENTICATION_SERVICES
-} from "graphql-module/queries";
-import { Loading } from "use-patternfly";
-import { dnsSubDomainRfc1123NameRegexp } from "types/Configs";
-
-import { StyleSheet, css } from "@patternfly/react-styles";
+import { css, StyleSheet } from "@patternfly/react-styles";
 
 export const dropdown_item_styles = StyleSheet.create({
   format_item: { whiteSpace: "normal", textAlign: "justify" }
 });
 
-export interface IAddressSpaceConfiguration {
-  name: string;
-  setName: (name: string) => void;
-  type: string;
-  setType: (type: string) => void;
-  plan: string;
-  setPlan: (plan: string) => void;
+export interface IAddressSpaceConfigurationProps {
+  onNameSpaceSelect: (event: any) => void;
+  setIsNameSpaceOpen: (isOpen: boolean) => void;
+  handleNameChange: (name: string) => void;
+  handleBrokeredChange: () => void;
+  onPlanSelect: (event: any) => void;
+  handleStandardChange: () => void;
+  setIsPlanOpen: (isOpen: boolean) => void;
+  onAuthenticationServiceSelect: (event: any) => void;
+  setIsAuthenticationServiceOpen: (isOpen: boolean) => void;
+  isNameSpaceOpen: boolean;
   namespace: string;
-  setNamespace: (namespace: string) => void;
-  authenticationService: string;
-  setAuthenticationService: (authenticationService: string) => void;
+  namespaceOptions: IDropdownOption[];
+  name: string;
   isNameValid: boolean;
-  setIsNameValid: (isNameValid: boolean) => void;
-}
-export interface IAddressSpacePlans {
-  addressSpacePlans: Array<{
-    metadata: {
-      name: string;
-      uid: string;
-      creationTimestamp: Date;
-    };
-    spec: {
-      addressSpaceType: string;
-      displayName: string;
-      longDescription: string;
-      shortDescription: string;
-    };
-  }>;
+  isStandardChecked: boolean;
+  isBrokeredChecked: boolean;
+  isPlanOpen: boolean;
+  type: string;
+  plan: string;
+  planOptions: any[];
+  isAuthenticationServiceOpen: boolean;
+  authenticationService: string;
+  authenticationServiceOptions: any[];
 }
 
-export interface IAddressSpaceAuthServiceResponse {
-  addressSpaceSchema_v2: IAddressSpaceAuthService[];
-}
-
-export interface IAddressSpaceAuthService {
-  metadata: {
-    name: string;
-  };
-  spec: {
-    authenticationServices: string[];
-  };
-}
-
-export interface INamespaces {
-  namespaces: Array<{
-    metadata: {
-      name: string;
-    };
-    status: {
-      phase: string;
-    };
-  }>;
-}
-
-export const AddressSpaceConfiguration: React.FunctionComponent<IAddressSpaceConfiguration> = ({
-  name,
-  setName,
+export const AddressSpaceConfiguration: React.FC<IAddressSpaceConfigurationProps> = ({
+  onNameSpaceSelect,
+  isNameSpaceOpen,
+  setIsNameSpaceOpen,
   namespace,
-  setNamespace,
-  type,
-  setType,
-  plan,
-  setPlan,
-  authenticationService,
-  setAuthenticationService,
+  namespaceOptions,
+  name,
   isNameValid,
-  setIsNameValid
+  handleNameChange,
+  handleStandardChange,
+  isBrokeredChecked,
+  handleBrokeredChange,
+  onPlanSelect,
+  isPlanOpen,
+  type,
+  setIsPlanOpen,
+  plan,
+  planOptions,
+  onAuthenticationServiceSelect,
+  isAuthenticationServiceOpen,
+  setIsAuthenticationServiceOpen,
+  authenticationService,
+  authenticationServiceOptions,
+  isStandardChecked
 }) => {
-  //TODO: Fix namespace value on the textbox
-  const [isNameSpaceOpen, setIsNameSpaceOpen] = React.useState(false);
-  const [isStandardChecked, setIsStandardChecked] = React.useState(false);
-  const [isBrokeredChecked, setIsBrokeredChecked] = React.useState(false);
-  const onNameSpaceSelect = (event: any) => {
-    event.currentTarget.childNodes[0] &&
-      setNamespace(event.currentTarget.childNodes[0].value);
-    setIsNameSpaceOpen(!isNameSpaceOpen);
-  };
-  const [isPlanOpen, setIsPlanOpen] = React.useState(false);
-  const onPlanSelect = (event: any) => {
-    event.currentTarget.childNodes[0] &&
-      setPlan(event.currentTarget.childNodes[0].value);
-    setIsPlanOpen(!isPlanOpen);
-  };
-
-  const [
-    isAuthenticationServiceOpen,
-    setIsAuthenticationServiceOpen
-  ] = React.useState(false);
-  const onAuthenticationServiceSelect = (event: any) => {
-    event.currentTarget.childNodes[0] &&
-      setAuthenticationService(event.currentTarget.childNodes[0].value);
-    setIsAuthenticationServiceOpen(!isAuthenticationServiceOpen);
-  };
-
-  React.useEffect(() => {
-    if (type === "standard") setIsStandardChecked(true);
-    else if (type === "brokered") setIsBrokeredChecked(true);
-  }, []);
-
-  const { loading, data } = useQuery<INamespaces>(RETURN_NAMESPACES);
-
-  const { data: authenticationServices } = useQuery<
-    IAddressSpaceAuthServiceResponse
-  >(RETURN_AUTHENTICATION_SERVICES) || { data: { addressSpaceSchema_v2: [] } };
-
-  const { addressSpacePlans } = useQuery<IAddressSpacePlans>(
-    RETURN_ADDRESS_SPACE_PLANS
-  ).data || {
-    addressSpacePlans: []
-  };
-
-  if (loading) return <Loading />;
-
-  const { namespaces } = data || {
-    namespaces: []
-  };
-
-  let namespaceOptions: IDropdownOption[];
-
-  let planOptions: any[] = [];
-
-  let authenticationServiceOptions: any[] = [];
-
-  namespaceOptions = namespaces.map(namespace => {
-    return {
-      value: namespace.metadata.name,
-      label: namespace.metadata.name
-    };
-  });
-  if (type) {
-    planOptions =
-      addressSpacePlans
-        .map(plan => {
-          if (plan.spec.addressSpaceType === type) {
-            return {
-              value: plan.metadata.name,
-              label: plan.spec.displayName || plan.metadata.name,
-              description:
-                plan.spec.shortDescription || plan.spec.longDescription
-            };
-          }
-        })
-        .filter(plan => plan !== undefined) || [];
-  }
-
-  if (authenticationServices) {
-    authenticationServices.addressSpaceSchema_v2.forEach(authService => {
-      if (authService.metadata.name === type) {
-        authenticationServiceOptions = authService.spec.authenticationServices.map(
-          service => {
-            return {
-              value: service,
-              label: service
-            };
-          }
-        );
-      }
-    });
-  }
-
-  const handleBrokeredChange = () => {
-    setIsBrokeredChecked(true);
-    setIsStandardChecked(false);
-    setPlan(" ");
-    setAuthenticationService(" ");
-    setType("brokered");
-  };
-
-  const handleStandardChange = () => {
-    setIsStandardChecked(true);
-    setIsBrokeredChecked(false);
-    setPlan(" ");
-    setAuthenticationService(" ");
-    setType("standard");
-  };
-
-  const handleNameChange = (name: string) => {
-    setName(name);
-    !dnsSubDomainRfc1123NameRegexp.test(name)
-      ? setIsNameValid(false)
-      : setIsNameValid(true);
-  };
-
   return (
     <>
       <Grid>
@@ -316,7 +175,7 @@ export const AddressSpaceConfiguration: React.FunctionComponent<IAddressSpaceCon
                     {plan}
                   </DropdownToggle>
                 }
-                dropdownItems={planOptions.map(option => (
+                dropdownItems={planOptions.map((option: any) => (
                   <DropdownItem
                     id={`cas-dropdown-item-plan${option.value}`}
                     key={option.value}
