+++ console/console-init/ui/src/modules/address-space/components/EditAddressSpace/EditAddressSpace.tsx
@@ -3,12 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-import React, { useState } from "react";
-import { useQuery } from "@apollo/react-hooks";
-import {
-  RETURN_ADDRESS_SPACE_PLANS,
-  RETURN_FILTERED_AUTHENTICATION_SERVICES
-} from "graphql-module/queries";
+import React from "react";
 import {
   Form,
   TextContent,
@@ -22,121 +17,27 @@ import {
   Button,
   Modal
 } from "@patternfly/react-core";
-import {
-  IAddressSpacePlans,
-  IAddressSpaceAuthServiceResponse
-} from "modules/address-space/dialogs/CreateAddressSpace/CreateAddressSpaceConfiguration";
-import { Loading } from "use-patternfly";
-import { useStoreContext, types } from "context-state-reducer";
-import { useMutationQuery } from "hooks";
-import { EDIT_ADDRESS_SPACE } from "graphql-module/queries";
+import { IAddressSpace } from "modules/address-space";
 
-export const EditAddressSpace: React.FunctionComponent<{}> = () => {
-  const { state, dispatch } = useStoreContext();
-  const { modalProps } = (state && state.modal) || {};
-  const { onConfirm, onClose } = modalProps || {};
+export interface IEditAddressSpaceProps {
+  onCloseDialog: () => void;
+  onConfirmDialog: () => void;
+  onPlanChange: (plan: string) => void;
+  onAuthServiceChange: (authservice: string) => void;
+  authServiceOptions: any[];
+  planOptions: any[];
+  addressSpace: IAddressSpace;
+}
 
-  const [addressSpace, setAddressSpace] = useState(modalProps.addressSpace);
-  const refetchQueries: string[] = ["all_address_spaces"];
-  const [setEditAddressSpaceQueryVariables] = useMutationQuery(
-    EDIT_ADDRESS_SPACE,
-    refetchQueries
-  );
-
-  const { loading, data } = useQuery<IAddressSpacePlans>(
-    RETURN_ADDRESS_SPACE_PLANS
-  );
-
-  const authServices = useQuery<IAddressSpaceAuthServiceResponse>(
-    RETURN_FILTERED_AUTHENTICATION_SERVICES,
-    {
-      variables: {
-        t: addressSpace.type
-      }
-    }
-  ).data || { addressSpaceSchema_v2: [] };
-
-  if (loading) return <Loading />;
-
-  const { addressSpacePlans } = data || {
-    addressSpacePlans: []
-  };
-
-  let planOptions: any[] = [];
-
-  let authServiceOptions: any[] = [];
-
-  if (addressSpace.type) {
-    planOptions =
-      addressSpacePlans
-        .map(plan => {
-          if (plan.spec.addressSpaceType === addressSpace.type) {
-            return {
-              value: plan.metadata.name,
-              label: plan.metadata.name
-            };
-          }
-        })
-        .filter(plan => plan !== undefined) || [];
-  }
-
-  if (authServices.addressSpaceSchema_v2[0])
-    authServiceOptions = authServices.addressSpaceSchema_v2[0].spec.authenticationServices.map(
-      authService => {
-        return {
-          value: authService,
-          label: authService
-        };
-      }
-    );
-
-  const onCloseDialog = () => {
-    dispatch({ type: types.HIDE_MODAL });
-    if (onClose) {
-      onClose();
-    }
-  };
-
-  const onPlanChange = (plan: string) => {
-    if (addressSpace) {
-      addressSpace.planValue = plan;
-      setAddressSpace({ ...addressSpace });
-    }
-  };
-
-  const onAuthServiceChanged = (authService: string) => {
-    if (addressSpace) {
-      addressSpace.authenticationService = authService;
-      setAddressSpace({ ...addressSpace });
-    }
-  };
-
-  const onConfirmDialog = async () => {
-    if (addressSpace) {
-      const variables = {
-        a: {
-          name: addressSpace.name,
-          namespace: addressSpace.nameSpace
-        },
-        jsonPatch:
-          '[{"op":"replace","path":"/spec/plan","value":"' +
-          addressSpace.planValue +
-          '"},' +
-          '{"op":"replace","path":"/spec/authenticationService/name","value":"' +
-          addressSpace.authenticationService +
-          '"}' +
-          "]",
-        patchType: "application/json-patch+json"
-      };
-      await setEditAddressSpaceQueryVariables(variables);
-
-      onCloseDialog();
-      if (onConfirm) {
-        onConfirm();
-      }
-    }
-  };
-
+export const EditAddressSpace: React.FC<IEditAddressSpaceProps> = ({
+  onCloseDialog,
+  onConfirmDialog,
+  onPlanChange,
+  onAuthServiceChange,
+  authServiceOptions,
+  planOptions,
+  addressSpace
+}) => {
   return (
     <Modal
       isLarge
@@ -241,7 +142,7 @@ export const EditAddressSpace: React.FunctionComponent<{}> = () => {
           <FormSelect
             id="edit-addr-auth"
             value={addressSpace.authenticationService}
-            onChange={val => onAuthServiceChanged(val)}
+            onChange={val => onAuthServiceChange(val)}
             aria-label="FormSelect Input"
           >
             {authServiceOptions.map((option, index) => (
