+++ console/console-init/ui/src/modules/address-space/components/index.ts
@@ -6,3 +6,6 @@ export * from "./AddressSpaceToolbar";
 export * from "./EmptyAddressSpace";
 export * from "./AddressSpaceFilter";
 export * from "./AddressSpaceList";
+export * from "./AddressSpaceConfiguration";
+export * from "./EditAddressSpace";
+export * from "./AddressSpaceReview";
