+++ console/console-init/ui/src/modules/address-space/dialogs/CreateAddressSpace/index.ts
@@ -4,5 +4,5 @@
  */
 
 export * from "./CreateAddressSpace";
-export * from "./CreateAddressSpaceConfiguration";
-export * from "./ReviewAddressSpace";
+export * from "./Configuration";
+export * from "./Review";
