+++ console/console-init/ui/src/modules/address-space/dialogs/index.ts
@@ -3,5 +3,5 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export * from "./EditAddressSpace";
+export * from "./EditAddressSpaceContainer";
 export * from "./CreateAddressSpace";
